Extract Select component
